hybridstress/requirements.txt
# HybridStress Benchmark Dependencies
# Install: pip install -r hybridstress/requirements.txt

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Vision backbone (CMV student)
open_clip_torch>=2.24.0
timm>=0.9.0

# VLM teacher (Qwen2-VL-7B)
transformers>=4.45.0
qwen-vl-utils>=0.0.8
accelerate>=0.26.0

# Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# OCR validators
pytesseract>=0.3.10
easyocr>=1.7.0

# Experiment tracking (optional)
wandb>=0.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
paramiko>=3.0.0
